Move module holding code for the extensions into the pydoctor.extensions package.
Use pydoctor.model.System by default.
